Cancer Dataset Used for Analysis
This is the raw dataset we used to conduct our project. We found the file on data.world at the following link: https://data.world/nrippner/ols-regression-challenge.  The data in this file was taken from the following websites by the creator of this dataset: census.gov, clinicaltrials.gov, and cancer.gov.